refactor(workspace): move timeline into workspace package
Move createTimeline, entry-types, and sheet CSV helpers from
packages/filesystem into packages/workspace/src/content/. Filesystem
now re-exports from workspace. Replace branded generateColumnId/
generateRowId with generateId() in sheet-csv.

Spec items: 1.1–1.6
